requirements.txt
fastapi
uvicorn
pymongo
requests
python-dotenv
passlib
jose
